pyproject.toml
[project]
name = "repl-nix-aiagent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.4",
    "langchain-community>=0.3.3",
    "langchain-openai>=0.2.4",
    "python-dotenv>=1.0.1",
    "langchain-core>=0.3.13",
    "google-search-results>=2.4.2",
]
